# contracts/openapi/safety_kernel.yaml
...
openapi: 3.1.0
info:
  title: QantmOrchstrtr-RSI Safety Kernel
  version: 1.0.0
  description: |
    Authorization, attestation, and audit-chain service for sensitive
    operations across the QantmOrchstrtr-RSI system. The kernel is the
    only component authorized to mint signed tokens for actions like
    `sio_run_cycles`, RSI cycle persistence, and approval decisions.

    Caller roles (set by middleware via `x-api-key`):
    - `worker` — primary caller; authorized to request tokens for
      RSI/SIO/training actions
    - `api` — REST API; restricted to governance-only actions per
      `is_api_action_allowed()` allowlist
    - `operator` — human approvals/rejections only
  contact:
    name: ARYA Labs Safety Kernel team
    url: https://github.com/ARYA-Labs-PBC/QantmOrchstrtr-RSI
  license:
    name: Proprietary
servers:
  - url: http://safety-kernel:9000
    description: Internal docker network (production)
  - url: http://localhost:9001
    description: Host port mapping (local development)
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: |
        Pre-shared key set in `.env` per role:
        - `QORCH_KERNEL_API_KEY_WORKER`
        - `QORCH_KERNEL_API_KEY_API`
        - `QORCH_KERNEL_API_KEY_OPERATOR`
  schemas:
...
    ErrorResponse:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          enum: [false]
        error:
          type: string
          description: Stable machine-readable error code
        reason:
          type: string
          description: Optional human-readable reason

    # ------------------------------------------------------------------
    # /policy/* (ADR-018, ARY-2028) — module-load / exec / compile
    # authorization. Slice 1 freezes ONLY PolicyModuleAuthorize{Request,
    # AllowResponse, DenyResponse}. The /policy/module/register,
    # /policy/audit-event, /policy/module/{path}/status endpoints are
    # listed in `paths:` with a DRAFT description and 501-only
    # responses; slice 2 fixes their request/response schemas.
    # ------------------------------------------------------------------
...
paths:
...
  /policy/module/register:
    post:
      operationId: policyModuleRegister
      summary: (DRAFT) Register a module path with required-pattern set
      description: |
        DRAFT — schema may change in slice 2; do not implement clients
        against this shape. Slice 2 will register a module path plus
        the regex pattern set required for future authorize calls on
        that path.
      tags: [policy]
      security:
        - ApiKeyAuth: []
      responses:
        '501':
          description: Not Implemented — slice-1 scaffold
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...